# backend/docs/openapi/schema.yml
openapi: 3.0.3
info:
  title: StudyAI API
  version: v1
paths:
...
  /api/v1/documents/{id}/enrich:
    post:
      operationId: api_v1_documents_enrich_create
      description: '§60: POST /documents/{id}/enrich — 200 existing / 202 enrich job.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Document'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/enrichment:
    get:
      operationId: api_v1_documents_enrichment_retrieve
      description: '§60: GET /documents/{id}/enrichment.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
...
  /api/v1/documents/{id}/pdf:
    post:
      operationId: api_v1_documents_pdf_create
      description: '§60: POST /documents/{id}/pdf — 200 existing artifact or 202 render
        job.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Document'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/refresh-ai:
    post:
      operationId: api_v1_documents_refresh_ai_create
      description: '§60: POST /documents/{id}/refresh-ai — forces regeneration.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Document'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/retry-processing:
    post:
      operationId: api_v1_documents_retry_processing_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Document'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/revisions:
    get:
      operationId: api_v1_documents_revisions_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
    post:
      operationId: api_v1_documents_revisions_create
      description: 'Two modes: finalize-upload (no lines) or user-edited revision
        (with lines).'
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Document'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
  /api/v1/documents/{id}/tags:
    get:
      operationId: api_v1_documents_tags_retrieve
      description: '§60: GET /documents/{id}/tags — stable tags linked to this document.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
...
  /api/v1/tags:
    get:
      operationId: api_v1_tags_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTagList'
          description: ''
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        profile:
          type: string
          format: uuid
          nullable: true
        subject:
          type: string
          format: uuid
          nullable: true
        source:
          $ref: '#/components/schemas/SourceEnum'
        source_type:
          $ref: '#/components/schemas/SourceTypeEnum'
        schema_version:
          type: string
          maxLength: 16
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - source
      - source_type
...
    PaginatedTagList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
    SourceEnum:
      enum:
      - upload
      - canvas
      - reference
      type: string
      description: |-
        * `upload` - upload
        * `canvas` - canvas
        * `reference` - reference
    SourceTypeEnum:
      enum:
      - image
      - pdf
      - canvas_page
      - reference
      type: string
      description: |-
        * `image` - image
        * `pdf` - pdf
        * `canvas_page` - canvas_page
        * `reference` - reference
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        subject:
          type: string
          format: uuid
        parent:
          type: string
          format: uuid
          nullable: true
        stable_key:
          type: string
          maxLength: 64
        display_name:
          type: string
          maxLength: 120
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - display_name
      - id
      - stable_key
      - subject
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT